Refactoring networking code
